+++ .gitignore
@@ -32,3 +32,4 @@ hs_err_pid*
 # JDT-specific (Eclipse Java Development Tools)     
 .classpath
 swaggen/generated/
+*.html
