+++ .travis.yml
@@ -16,7 +16,6 @@ jobs:
       before_script:
         - cd swaggen
       script:
-        - npm install -g swagger-cli
         - npm install -g redoc-cli
         - mvn -U -s ../assets/settings.xml clean install #plugin
         - mvn javadoc:javadoc -s ../assets/settings.xml
